rename non-conventionally named package
