+++ rpc/tendermint/test/client_rpc_test.go
@@ -1,4 +1,4 @@
-package rpctest
+package test
 
 import (
 	"testing"
