+++ rpc/tendermint/test/client_ws_test.go
@@ -1,4 +1,4 @@
-package rpctest
+package test
 
 import (
 	"fmt"
