+++ rpc/tendermint/test/common.go
@@ -1,4 +1,4 @@
-package rpctest
+package test
 
 import (
 	"fmt"
