+++ rpc/tendermint/test/common_test.go
@@ -1,4 +1,4 @@
-package rpctest
+package test
 
 import (
 	"os"
